requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pyairtable==2.0.0
python-dotenv==1.0.0
pydantic==2.3.0
python-multipart==0.0.6
